+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- AccountCreatePage will check shopper's account status and forbid account signup if status is "approved", "rejected", or "under-review".
+
 ## [0.0.11] - 2020-06-02
 
 ### Fixed